1. 字体宽度计算重置 2. 弹幕毫秒级处理

+++ src/main/java/skiree/host/danmu/model/engine/DanMuBar.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class DanMuBar {
     public String pos;
     public Boolean mark;
-    public double nTime;
+    public int nTime;
 
     public DanMuBar(String pos) {
         this.pos = pos;
